+++ lib/fs/filesystem.js
@@ -1,0 +1,357 @@
+const {
+  FileProxy, ACCESS_READ, ACCESS_WRITE, ACCESS_ALLOWED, ACCESS_FLAGS,
+} = require('./fileproxy');
+const { logger } = require('../rest/client');
+
+// nodejs fs module work-alike.
+class FileSystem {
+  constructor(rest) {
+    this.rest = rest;
+
+    this.fds = [];
+    this.statCache = {};
+  }
+
+  stat(path, callback) {
+    // Listing a directory fills this cache to speed up the all but certain
+    // stat() calls to follow.
+    const info = this.statCache[path];
+    if (info) {
+      delete this.statCache[path];
+      callback(null, info);
+      return;
+    }
+
+    this.rest.info(path, callback);
+  }
+
+  lstat(...args) {
+    return this.stat(...args);
+  }
+
+  unlink(path, callback) {
+    this.rest.delete(path, (e, json) => {
+      // Ignore a 404 when deleting, if path does not exist, then delete was
+      // a "success".
+      if (e && e.statusCode === 404
+          && e.options.uri === '/api/2/path/oper/remove/') {
+        // Mask out the error, and forge a fake response.
+        callback(null, {
+          result: {
+            status: 'SUCCESS',
+          },
+        });
+        return;
+      }
+
+      callback(e, json);
+    });
+  }
+
+  rmdir(...args) {
+    return this.unlink(...args);
+  }
+
+  mkdir(path, callback) {
+    this.rest.mkdir(path, callback);
+  }
+
+  copy(src, dst, callback) {
+    this.rest.copy(src, dst, callback);
+  }
+
+  copyFile(...args) {
+    return this.copy(...args);
+  }
+
+  copyDir(...args) {
+    return this.copy(...args);
+  }
+
+  move(src, dst, callback) {
+    this.rest.move(src, dst, callback);
+  }
+
+  rename(src, dst, callback) {
+    this.rest.rename(src, dst, callback);
+  }
+
+  exists(path, callback) {
+    this.stat(path, (e) => {
+      // TODO: check for 404.
+      if (e) {
+        callback(null, false);
+        return;
+      }
+
+      callback(null, true);
+    });
+  }
+
+  createReadStream(path) {
+    // TODO: options is optional.
+    // https://nodejs.org/api/fs.html#fs_fs_createreadstream_path_options
+    return this.rest.download(path);
+  }
+
+  // eslint-disable-next-line no-unused-vars, class-methods-use-this
+  createWriteStream(path, options) {
+    throw new Error('Not yet supported by API');
+    // TODO: options is optional.
+    // https://nodejs.org/api/fs.html#fs_fs_createwritestream_path_options
+    // return this.rest.upload(path);
+  }
+
+  open(path, _flags, _mode, _callback) {
+    // https://nodejs.org/api/fs.html#fs_fs_open_path_flags_mode_callback
+    let file = null;
+    let callback;
+
+    // Shift arguments.
+    if (typeof _mode === 'function') {
+      callback = _mode;
+    } else {
+      callback = _callback;
+    }
+
+    const flags = _flags || 'r';
+    const accessType = ACCESS_FLAGS[flags];
+
+    try {
+      switch (accessType) {
+        /*
+        Not currently any special handling for read or write.
+
+        case ACCESS_READ:
+        file = new FileReader(this, path, flags, callback);
+        break;
+
+        case ACCESS_WRITE:
+        file = new FileWriter(this, path, flags, callback);
+        break;
+        */
+
+        case ACCESS_READ:
+        case ACCESS_WRITE:
+        case ACCESS_ALLOWED:
+          file = new FileProxy(this, path, flags, accessType, callback);
+          break;
+
+        default:
+          throw new Error(`invalid flags: ${flags} (${accessType})`);
+      }
+    } catch (e) {
+      return callback(e);
+    }
+    return file;
+  }
+
+  close(fd, callback) {
+    const file = this.fds[fd];
+
+    if (!file) {
+      callback(new Error(`invalid fd ${fd}`));
+      return;
+    }
+
+    file.close(callback);
+  }
+
+  fstat(fd, callback) {
+    const file = this.fds[fd];
+
+    if (!file) {
+      callback(new Error(`invalid fd ${fd}`));
+      return;
+    }
+
+    this.stat(file.path, callback);
+  }
+
+  read(fd, buffer, offset, length, position, callback) {
+    // https://nodejs.org/api/fs.html#fs_fs_read_fd_buffer_offset_length_position_callback
+    const file = this.fds[fd];
+
+    if (!file) {
+      callback(new Error(`invalid fd ${fd}`));
+      return;
+    }
+
+    file.read(buffer, offset, length, position, callback);
+  }
+
+  readFile(path, _options, _callback) {
+    // https://nodejs.org/api/fs.html#fs_fs_readfile_path_options_callback
+    let buffer = Buffer.alloc();
+    let pos = 0;
+    let callback;
+
+    // Shift arguments.
+    if (typeof _options === 'function') {
+      callback = _options;
+    } else {
+      callback = _callback;
+    }
+
+    this.open(path, 'r', null, (openError, fd) => {
+      if (openError) {
+        callback(openError);
+        return;
+      }
+
+      function readChunk() {
+        const chunk = Buffer.alloc(16384);
+
+        this.read(fd, chunk, 0, 16384, pos, (readError, bytesRead) => {
+          if (readError) {
+            callback(readError);
+            return;
+          }
+          if (bytesRead === 0) {
+            callback(null, buffer);
+            return;
+          }
+
+          pos += bytesRead;
+          buffer = Buffer.concat(buffer, chunk);
+
+          setInterval(readChunk, 0);
+        });
+      }
+
+      setInterval(readChunk, 0);
+    });
+  }
+
+  write(fd, buffer, offset, length, position, callback) {
+    // https://nodejs.org/api/fs.html#fs_fs_write_fd_buffer_offset_length_position_callback
+    // https://nodejs.org/api/fs.html#fs_fs_write_fd_string_position_encoding_callback
+    const file = this.fds[fd];
+
+    if (!file) {
+      callback(new Error(`invalid fd ${fd}`));
+      return;
+    }
+
+    file.write(buffer, offset, length, position, callback);
+  }
+
+  writeFile(path, buffer, _options, _callback) {
+    // https://nodejs.org/api/fs.html#fs_fs_writefile_file_data_options_callback
+    let pos = 0;
+    let callback;
+
+    // Shift arguments.
+    if (typeof _options === 'function') {
+      callback = _options;
+    } else {
+      callback = _callback;
+    }
+
+    this.open(path, 'w', null, (openError, fd) => {
+      if (openError) {
+        callback(openError);
+        return;
+      }
+
+      function writeChunk() {
+        this.write(fd, buffer, pos, 16384, pos, (writeError, bytesWritten, chunk) => {
+          if (writeError) {
+            callback(writeError);
+            return;
+          }
+
+          pos += bytesWritten;
+          if (bytesWritten === chunk.byteLength) {
+            callback(null);
+            return;
+          }
+
+          setInterval(writeChunk, 0);
+        });
+      }
+
+      setInterval(writeChunk, 0);
+    });
+  }
+
+  // TODO: make this call readdirstats() and strip out just the names (while
+  // allowing it to populate statCache).
+  // TODO: add incremental option to call callback multiple times instead of
+  // buffering.
+  readdir(path, callback, options) {
+    // https://nodejs.org/api/fs.html#fs_fs_readdir_path_options_callback
+    this.readdirstats(path, (e, infos) => {
+      if (e) {
+        callback(e);
+        return;
+      }
+
+      let names = null;
+
+      if (infos) {
+        names = [];
+
+        for (let i = 0; i < infos.length; i++) {
+          names[i] = infos[i].name;
+        }
+      }
+
+      callback(null, names);
+    }, options);
+  }
+
+  readdirstats(path, callback, opts) {
+    // eslint-disable-next-line prefer-object-spread
+    const options = Object.assign({
+      incremental: false,
+    }, opts);
+
+    // API arguments.
+    const args = {
+      children: true,
+    };
+
+    // Buffer when not incremental.
+    const infos = options.incremental ? null : [];
+
+    // Clear cache, only meant to be used for stat() calls following a
+    // readdir().
+    this.statCache = {};
+
+    // Make API call.
+    this.rest.info(path, (e, page) => {
+      if (e) {
+        callback(e);
+        return;
+      }
+
+      // Last page, call callback.
+      if (page === null) {
+        // Not that infos is null when incremental.
+        callback(null, infos);
+        return;
+      }
+
+      if (options.incremental) {
+        callback(null, page);
+      }
+
+      // Buffer page of results.
+      for (let i = 0; i < page.length; i++) {
+        if (!options.incremental) {
+          infos.push(page[i]);
+        }
+
+        // Cache stat information, most times stat() calls for everything we
+        // list are imminent.
+        this.statCache[page[i].path] = page[i];
+      }
+    }, args);
+  }
+}
+
+module.exports = {
+  FileSystem,
+  logger,
+};
