+++ lib/rest/client.js
@@ -337,36 +337,8 @@ class Client {
   }
 }
 
-class BasicClient extends Client {
-  constructor(opts) {
-    // eslint-disable-next-line prefer-object-spread
-    const options = Object.assign({}, opts);
-
-    if (!options.auth) {
-      const username = options.username || process.env.SMARTFILE_USER;
-      const password = options.password || process.env.SMARTFILE_PASS;
-
-      delete options.username;
-      delete options.password;
-
-      if (!username || !password) {
-        throw new Error('username and password require for basic auth');
-      }
-
-      options.auth = {
-        user: username,
-        pass: password,
-        sendImmediately: true,
-      };
-    }
-
-    super(options);
-  }
-}
-
 module.exports = {
   Client,
-  BasicClient,
   encodePath,
   normPath,
   logger,
